+++ src/css.test.ts
@@ -19,7 +19,7 @@ describe('css', () => {
 				color: red;
 			`,
 		])
-		expect(Tag1.className).toMatchInlineSnapshot(
+		expect(Tag1.class).toMatchInlineSnapshot(
 			'"testFile__fspbdo1 testFile__fspbdo0"'
 		)
 		const out = css`